refactor:리스너랑 함수를 구분하고 state로 같이 쓰는 값을 정의
