+++ js/function/totalAmount.js
