+++ js/index.js
@@ -1,6 +1,7 @@
 // js/index.js
 import { initCalendar } from "./header.js";
 import { initInputForm } from "./input-form.js";
+import { initListener } from "./listener/totalListener.js"; 
 
 initCalendar({
   onUpdate: (year, month) => {
@@ -9,3 +10,4 @@ initCalendar({
 });
 
 initInputForm();
+initListener();
