+++ js/input-form.js
@@ -1,51 +1,21 @@
 //js/input-form.js
-import { updateTotalAmounts } from "./totalAmount.js";
-import { dummyEntries } from "./dummy.js";
+import { updateTotalAmounts } from "./function/totalAmount.js";
+import { dummyEntries } from "./dummy/dummy.js";
+import { sharedState } from "./state/state.js";
+import { renderCategoryOptions } from "./function/categoryRender.js";
 
 export function initInputForm() {
-  let selectedMethod = null;
-  let selectedCategory = null;
-  let isIncome = true;
+  let selectedMethod = sharedState.selectedMethod; // sharedState에서 selectedMethod 요소를 가져옴
+  let selectedCategory = sharedState.selectedCategory; // sharedState에서 selectedCategory 요소를 가져옴
+  let isIncome = sharedState.isIncome; // 수입/지출 여부를 sharedState에서 가져옴
 
-  const toggleSign = document.getElementById("toggle-sign");
   const addBtn = document.getElementById("add-btn");
 
-  const amountInput = document.getElementById("amount");
-  const descInput = document.getElementById("desc");
-  const charCount = document.getElementById("char-count");
-
   const display = document.getElementById("dropdown-display");
-  const panel = document.getElementById("dropdown-panel");
-  const dropAddBtn = document.getElementById("dropdown-add");
-
-  const modal = document.getElementById("modal");
-  const confirmAdd = document.getElementById("confirm-add");
-  const cancelAdd = document.getElementById("cancel-add");
-  const newMethodInput = document.getElementById("new-method");
-
-  const categoryWrapper = document.getElementById("category-wrapper");
   const categoryDisplay = document.getElementById("category-display");
-  const categoryPanel = document.getElementById("category-panel");
 
-  const incomeCategories = ["월급", "용돈", "기타수입"];
-  const expenseCategories = ["생활", "식비", "교통", "쇼핑/뷰티", "의료/건강", "문화/여가", "미분류"];
 
   const entries = [];
-
-  function renderCategoryOptions() {
-    // 수입/지출에 따라 카테고리 옵션을 업데이트
-    updateTotalAmounts();
-    categoryPanel.innerHTML = "";
-    const list = isIncome ? incomeCategories : expenseCategories;
-
-    list.forEach(item => {
-      const div = document.createElement("div");
-      div.className = "dropdown-option";
-      div.textContent = item;
-      div.dataset.value = item;
-      categoryPanel.appendChild(div);
-    });
-  }
 
   function addEntryToDOM(entry) {
     const entryList = document.getElementById("entry-list");
@@ -87,97 +57,6 @@ export function initInputForm() {
     `;
     entryItems.appendChild(item);
   }
-
-  toggleSign.addEventListener("click", () => {
-    isIncome = !isIncome;
-    toggleSign.textContent = isIncome ? "+" : "-";
-    toggleSign.classList.toggle("minus", !isIncome);
-    categoryDisplay.textContent = "선택하세요";
-    selectedCategory = null;
-    renderCategoryOptions();
-  });
-
-  amountInput.addEventListener("input", () => {
-    const rawValue = amountInput.value.replace(/[^\d]/g, "");
-    amountInput.value = rawValue ? Number(rawValue).toLocaleString() : "";
-  });
-
-  descInput.addEventListener("input", () => {
-    const length = descInput.value.length;
-    charCount.textContent = `${length} / 32`;
-  });
-
-  display.addEventListener("click", () => {
-    panel.classList.toggle("hidden");
-  });
-
-  dropAddBtn.addEventListener("click", () => {
-    modal.classList.remove("hidden");
-    newMethodInput.value = "";
-    panel.classList.add("hidden");
-  });
-
-  cancelAdd.addEventListener("click", () => {
-    modal.classList.add("hidden");
-  });
-
-  confirmAdd.addEventListener("click", () => {
-    const newMethod = newMethodInput.value.trim();
-    if (!newMethod) {
-      alert("결제수단을 입력해주세요.");
-      return;
-    }
-
-    const option = document.createElement("div");
-    option.className = "dropdown-option";
-    option.dataset.value = newMethod;
-    option.innerHTML = `<span>${newMethod}</span><button class="delete-btn">❌</button>`;
-    panel.insertBefore(option, dropAddBtn);
-
-    display.textContent = newMethod;
-    selectedMethod = newMethod;
-    modal.classList.add("hidden");
-  });
-
-  document.addEventListener("click", (e) => {
-    if (!document.getElementById("method-wrapper").contains(e.target)) {
-      panel.classList.add("hidden");
-    }
-    if (!categoryWrapper.contains(e.target)) {
-      categoryPanel.classList.add("hidden");
-    }
-  });
-
-  panel.addEventListener("click", (e) => {
-    const optionDiv = e.target.closest(".dropdown-option");
-    if (!optionDiv) return;
-
-    const methodName = optionDiv.dataset.value;
-
-    if (e.target.classList.contains("delete-btn")) {
-      optionDiv.remove();
-      if (selectedMethod === methodName) {
-        display.textContent = "선택하세요";
-        selectedMethod = null;
-      }
-    } else {
-      display.textContent = methodName;
-      selectedMethod = methodName;
-      panel.classList.add("hidden");
-    }
-  });
-
-  categoryDisplay.addEventListener("click", () => {
-    categoryPanel.classList.toggle("hidden");
-  });
-
-  categoryPanel.addEventListener("click", (e) => {
-    if (e.target.classList.contains("dropdown-option")) {
-      selectedCategory = e.target.dataset.value;
-      categoryDisplay.textContent = selectedCategory;
-      categoryPanel.classList.add("hidden");
-    }
-  });
 
   addBtn.addEventListener("click", () => {
     const date = document.getElementById("date").value;
